@@ -11,43 +11,54 @@ description: Change the world.
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.1.0+11
+version: 1.3.0-alpha.2+18
 
 environment:
-  sdk: ">=2.13.0 <3.0.0"
+  sdk: ">=2.14.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  flutter_svg: ^1.0.0
+  flutter_svg: ^1.0.3
   equatable: ^2.0.3
   flutter_bloc: ^8.0.1
   get_it: ^7.2.0
-  image: ^3.1.0
-  #package_info_plus: ^1.3.0
+  image: ^3.1.3
   intl: ^0.17.0
-  path_provider: ^2.0.8
-  shared_preferences: ^2.0.11
-  url_launcher: ^6.0.17
-  file_picker: ^4.3.0
+  path_provider: ^2.0.9
+  shared_preferences: ^2.0.13
+  url_launcher: ^6.0.20
+  file_picker: ^4.5.1
   phosphor_flutter: ^1.4.0
   replay_bloc: ^0.2.1
-  share_plus: ^3.0.4
-  package_info_plus: ^1.3.0
+  share_plus: ^4.0.4
+  package_info_plus: ^1.4.2
   idb_shim: ^2.0.1
-  bitsdojo_window: ^0.1.1+1
   flutter_launcher_icons: ^0.9.2
-  go_router: ^2.5.7
+  go_router: ^3.0.7
   animations: ^2.0.2
   xml: ^5.3.1
+  collection: ^1.16.0
+  bloc_concurrency: ^0.2.0
+  camera: ^0.9.4+19
+  camera_windows: ^0.1.0
+  window_manager: ^0.2.1
+  flex_color_scheme: ^4.2.0
+  freezed_annotation: ^1.1.0
+  json_annotation: ^4.4.0
+  path: ^1.8.1
 
 dev_dependencies:
-  flutter_native_splash: ^1.3.3
-  flutter_lints: ^1.0.4
+  flutter_native_splash: ^2.1.6
+  flutter_lints: ^2.0.0
   flutter_test:
     sdk: flutter
+  analyzer: ^3.4.1
+  freezed: ^1.1.1
+  build_runner: ^2.1.10
+  json_serializable: ^6.1.6
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
