@@ -3,9 +3,12 @@ import 'dart:convert';
 import 'package:butterfly/api/file_system.dart';
 import 'package:butterfly/bloc/document_bloc.dart';
 import 'package:butterfly/dialogs/import.dart';
-import 'package:butterfly/models/document.dart';
+import 'package:butterfly/models/viewport.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+
+import '../models/converter.dart';
+import '../renderers/renderer.dart';
 
 class ImportIntent extends Intent {
   final BuildContext context;
@@ -23,9 +26,19 @@ class ImportAction extends Action<ImportIntent> {
             builder: (context) => const ImportDialog(), context: intent.context)
         .then((content) {
       if (content == null) return;
-      var document = AppDocument.fromJson(jsonDecode(content));
-      DocumentFileSystem.fromPlatform().importDocument(document).then((file) {
-        bloc.emit(DocumentLoadSuccess(document, path: file.path));
+      var document = const DocumentJsonConverter()
+          .fromJson(Map<String, dynamic>.from(jsonDecode(content)));
+      DocumentFileSystem.fromPlatform()
+          .importDocument(document)
+          .then((file) async {
+        final background = Renderer.fromInstance(document.background);
+        await background.setup(document);
+        final renderers =
+            document.content.map((e) => Renderer.fromInstance(e)).toList();
+        await Future.wait(renderers.map((e) async => await e.setup(document)));
+        bloc.emit(DocumentLoadSuccess(document,
+            path: file.path,
+            cameraViewport: CameraViewport.unbaked(background, renderers)));
       });
     });
   }
