@@ -1,10 +1,11 @@
 import 'package:butterfly/api/file_system.dart';
 import 'package:butterfly/bloc/document_bloc.dart';
-import 'package:butterfly/dialogs/file_system.dart';
 import 'package:butterfly/models/document.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+
+import '../dialogs/file_system/dialog.dart';
 
 class OpenIntent extends Intent {
   final BuildContext context;
