@@ -1,6 +1,4 @@
-import 'package:butterfly/bloc/document_bloc.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 
 class RedoIntent extends Intent {
   final BuildContext context;
@@ -13,6 +11,7 @@ class RedoAction extends Action<RedoIntent> {
 
   @override
   void invoke(RedoIntent intent) {
-    intent.context.read<DocumentBloc>().redo();
+    // reimplement undo
+    // intent.context.read<DocumentBloc>().redo();
   }
 }
