@@ -1,6 +1,4 @@
-import 'package:butterfly/bloc/document_bloc.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 
 class UndoIntent extends Intent {
   final BuildContext context;
@@ -13,6 +11,7 @@ class UndoAction extends Action<UndoIntent> {
 
   @override
   void invoke(UndoIntent intent) {
-    intent.context.read<DocumentBloc>().undo();
+    // reimplement undo
+    //intent.context.read<DocumentBloc>().undo();
   }
 }
