@@ -4,10 +4,37 @@ import 'package:butterfly/api/file_system_io.dart';
 import 'package:butterfly/api/file_system_web.dart';
 import 'package:butterfly/models/document.dart';
 import 'package:butterfly/models/palette.dart';
+import 'package:butterfly/models/template.dart';
+import 'package:flutter/widgets.dart';
 import 'package:flutter/foundation.dart';
 import 'package:path/path.dart' as path;
 
-abstract class DocumentFileSystem {
+abstract class GeneralFileSystem {
+  String convertNameToFile(String name) {
+    return '${name.replaceAll(RegExp(r'[^a-zA-Z0-9]'), '_')}.bfly';
+  }
+
+  FutureOr<String> getAbsolutePath(String relativePath) async {
+    // Convert \ to /
+    relativePath = relativePath.replaceAll('\\', '/');
+    // Remove leading slash
+    if (relativePath.startsWith('/')) {
+      relativePath = relativePath.substring(1);
+    }
+    final root = await getDirectory();
+    var absolutePath = path.join(root, relativePath);
+    if (!absolutePath.startsWith(root)) {
+      throw Exception('Path is not in root directory');
+    }
+    return absolutePath;
+  }
+
+  FutureOr<String> getDirectory() {
+    return '/';
+  }
+}
+
+abstract class DocumentFileSystem extends GeneralFileSystem {
   Future<AppDocumentDirectory> getRootDirectory() {
     return getAsset('').then((value) => value as AppDocumentDirectory);
   }
@@ -64,29 +91,6 @@ abstract class DocumentFileSystem {
     }
   }
 
-  String convertNameToFile(String name) {
-    return name.replaceAll(RegExp(r'[^a-zA-Z0-9]'), '_') + '.bfly';
-  }
-
-  FutureOr<String> getAbsolutePath(String relativePath) async {
-    // Convert \ to /
-    relativePath = relativePath.replaceAll('\\', '/');
-    // Remove leading slash
-    if (relativePath.startsWith('/')) {
-      relativePath = relativePath.substring(1);
-    }
-    final root = await getDirectory();
-    var absolutePath = path.join(root, relativePath);
-    if (!absolutePath.startsWith(root)) {
-      throw Exception('Path is not in root directory');
-    }
-    return absolutePath;
-  }
-
-  FutureOr<String> getDirectory() {
-    return '/';
-  }
-
   Future<AppDocumentAsset?> duplicateAsset(String path, String newPath) async {
     var asset = await getAsset(path);
     if (asset == null) return null;
@@ -110,5 +114,50 @@ abstract class DocumentFileSystem {
     if (asset == null) return null;
     if (path != newPath) await deleteAsset(path);
     return asset;
+  }
+}
+
+abstract class TemplateFileSystem extends GeneralFileSystem {
+  Future<bool> createDefault(BuildContext context, {bool force = false});
+
+  Future<DocumentTemplate?> getTemplate(String name);
+  Future<DocumentTemplate> createTemplate(AppDocument document) async {
+    var template = DocumentTemplate(document: document);
+    var name = document.name;
+    var attemps = 1;
+    while (await hasTemplate(name)) {
+      name = '${document.name} ($attemps)';
+      attemps++;
+    }
+    template =
+        template.copyWith(document: template.document.copyWith(name: name));
+    updateTemplate(template);
+    return template;
+  }
+
+  Future<bool> hasTemplate(String name);
+  Future<void> updateTemplate(DocumentTemplate template);
+  Future<void> deleteTemplate(String name);
+  Future<List<DocumentTemplate>> getTemplates();
+
+  Future<DocumentTemplate?> renameTemplate(String path, String newName) async {
+    // Remove leading slash
+    if (path.startsWith('/')) {
+      path = path.substring(1);
+    }
+    final template = await getTemplate(path);
+    if (template == null) return null;
+    DocumentTemplate? newTemplate =
+        await createTemplate(template.document.copyWith(name: newName));
+    await deleteTemplate(path);
+    return newTemplate;
+  }
+
+  static TemplateFileSystem fromPlatform() {
+    if (kIsWeb) {
+      return WebTemplateFileSystem();
+    } else {
+      return IOTemplateFileSystem();
+    }
   }
 }
