@@ -4,5 +4,5 @@ Future<bool> openHelp(List<String> pageLocation) {
   return launch(Uri(
       scheme: 'https',
       host: 'docs.butterfly.linwood.dev',
-      pathSegments: ['docs', '1.1', ...pageLocation]).toString());
+      pathSegments: ['docs', '1.3', ...pageLocation]).toString());
 }
