@@ -5,7 +5,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:phosphor_flutter/phosphor_flutter.dart';
 
-import '../../models/painters/layer.dart';
+import '../../models/painter.dart';
 
 class LayerPainterDialog extends StatefulWidget {
   final DocumentBloc bloc;
@@ -27,8 +27,8 @@ class _LayerPainterDialogState extends State<LayerPainterDialog> {
   Widget build(BuildContext context) {
     return BlocProvider.value(
       value: widget.bloc,
-      child: Dialog(child:
-          BlocBuilder<DocumentBloc, DocumentState>(builder: (context, state) {
+      child: Dialog(child: Builder(builder: (context) {
+        var state = widget.bloc.state;
         if (state is! DocumentLoadSuccess) return Container();
         var painter =
             state.document.painters[widget.painterIndex] as LayerPainter;
@@ -85,7 +85,9 @@ class _LayerPainterDialogState extends State<LayerPainterDialog> {
                                 title: Text(AppLocalizations.of(context)!
                                     .includeEraser),
                                 onChanged: (value) => setState(() => painter =
-                                    painter.copyWith(includeEraser: value))),
+                                    painter.copyWith(
+                                        includeEraser:
+                                            value ?? painter.includeEraser))),
                           ]),
                         ),
                         const Divider(),
