@@ -1,59 +1,45 @@
 import 'dart:ui' as ui;
 
-import 'package:butterfly/models/backgrounds/box.dart';
+import 'package:butterfly/models/viewport.dart';
 import 'package:butterfly/models/document.dart';
-import 'package:butterfly/models/elements/element.dart';
-import 'package:butterfly/models/elements/image.dart';
+import 'package:butterfly/models/element.dart';
+import 'package:butterfly/renderers/renderer.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 import 'cubits/transform.dart';
+import 'models/area.dart';
 
 Future<ui.Image> loadImage(ImageElement layer) {
   return decodeImageFromList(layer.pixels);
 }
 
-void paintElement(Canvas canvas, PadElement element,
-    [Map<PadElement, ui.Image> images = const {}, bool preview = false]) {
-  if (element is ImageElement) {
-    if (images.containsKey(element)) {
-      // Resize image to scale of the element
-      // Get image from element pixels
-      var image = images[element];
-      if (image == null) return;
-      var scale = element.scale;
-      var width = image.width * scale;
-      var height = image.height * scale;
-      var paint = Paint()..isAntiAlias = true;
-      canvas.drawImageRect(
-        image,
-        Rect.fromLTWH(0, 0, image.width.toDouble(), image.height.toDouble()),
-        Rect.fromLTWH(element.position.dx, element.position.dy, width, height),
-        paint,
-      );
-    }
-  } else {
-    element.paint(canvas, preview);
-  }
-}
-
 class ForegroundPainter extends CustomPainter {
-  final Map<int, PadElement> editingLayer;
+  final List<Renderer> renderers;
   final CameraTransform transform;
-  final PadElement? selection;
+  final List<Rect> selection;
 
-  ForegroundPainter(this.editingLayer,
-      [this.transform = const CameraTransform(), this.selection]);
+  ForegroundPainter(
+    this.renderers, [
+    this.transform = const CameraTransform(),
+    this.selection = const [],
+  ]);
 
   @override
   void paint(Canvas canvas, Size size) {
     canvas.scale(transform.size);
     canvas.translate(transform.position.dx, transform.position.dy);
-    for (var element in editingLayer.entries) {
-      paintElement(canvas, element.value, {}, true);
+    for (var element in renderers) {
+      element.build(canvas, size, transform, true);
     }
-    if (selection != null) {
-      var rect = selection!.rect;
+    for (var rect in selection) {
+      /*
+      final minX =
+          -transform.position.dx + 20 / ((transform.size - 1) / 1.5 + 1);
+      final maxX = minX + size.width / transform.size - 40 / transform.size;
+      final minY = -transform.position.dy + 20;
+      final maxY = minY + size.height / transform.size - 40 / transform.size;
+      */
       canvas.drawRRect(
           RRect.fromRectAndRadius(rect.inflate(5), const Radius.circular(5)),
           Paint()
@@ -63,13 +49,13 @@ class ForegroundPainter extends CustomPainter {
               rect.bottomRight,
               [Colors.red, Colors.yellow],
             )
-            ..strokeWidth = 5);
+            ..strokeWidth = 10 / transform.size);
     }
   }
 
   @override
   bool shouldRepaint(ForegroundPainter oldDelegate) =>
-      oldDelegate.editingLayer != editingLayer ||
+      oldDelegate.renderers != renderers ||
       oldDelegate.transform != transform ||
       oldDelegate.selection != selection;
 }
@@ -77,9 +63,7 @@ class ForegroundPainter extends CustomPainter {
 Future<Map<PadElement, ui.Image>> loadImages(AppDocument document,
     [Map<PadElement, ui.Image> loadedImages = const {}]) async {
   var images = Map<PadElement, ui.Image>.from(loadedImages);
-  if (kIsWeb && document.content.any((element) => element is ImageElement)) {
-    await Future.delayed(const Duration(seconds: 1));
-  }
+  if (kIsWeb && document.content.any((element) => element is ImageElement)) {}
   for (var layer in document.content) {
     if (layer is ImageElement && !images.containsKey(layer)) {
       images[layer] = await loadImage(layer);
@@ -91,70 +75,64 @@ Future<Map<PadElement, ui.Image>> loadImages(AppDocument document,
 
 class ViewPainter extends CustomPainter {
   final AppDocument document;
+  final Area? currentArea;
   final bool renderBackground;
-  final List<String> invisibleLayers;
-  final Map<PadElement, ui.Image> images;
+  final CameraViewport cameraViewport;
   final CameraTransform transform;
 
-  ViewPainter(this.document,
-      {this.renderBackground = true,
-      this.invisibleLayers = const [],
-      this.transform = const CameraTransform(),
-      Map<PadElement, ui.Image>? images})
-      : images = images ?? <PadElement, ui.Image>{};
+  ViewPainter(
+    this.document, {
+    this.currentArea,
+    this.renderBackground = true,
+    required this.cameraViewport,
+    this.transform = const CameraTransform(),
+  });
 
   @override
   void paint(Canvas canvas, Size size) {
-    var background = document.background;
-    if (background is BoxBackground && renderBackground) {
-      canvas.drawColor(background.boxColor, BlendMode.srcOver);
-      if (background.boxWidth > 0 && background.boxXCount > 0) {
-        double x =
-            (transform.position.dx % background.boxWidth * transform.size);
-        int count = 0;
-        while (x < size.width) {
-          canvas.drawLine(
-              Offset(x, 0),
-              Offset(x, size.height),
-              Paint()
-                ..strokeWidth = background.boxXStroke
-                ..color = background.boxXColor);
-          count++;
-          if (count >= background.boxXCount) {
-            count = 0;
-            x += background.boxXSpace;
-          }
-          x += background.boxWidth * transform.size;
-        }
-      }
-      if (background.boxHeight > 0 && background.boxYCount > 0) {
-        double y =
-            (transform.position.dy % background.boxHeight * transform.size);
-        int count = 0;
-        while (y < size.width) {
-          canvas.drawLine(
-              Offset(0, y),
-              Offset(size.width, y),
-              Paint()
-                ..strokeWidth = background.boxYStroke
-                ..color = background.boxYColor);
-          count++;
-          if (count >= background.boxYCount) {
-            count = 0;
-            y += background.boxYSpace * transform.size;
-          }
-          y += background.boxHeight * transform.size;
-        }
+    var areaRect = currentArea?.rect;
+    if (areaRect != null) {
+      areaRect = Rect.fromPoints(transform.globalToLocal(areaRect.topLeft),
+          transform.globalToLocal(areaRect.bottomRight));
+    }
+    if (areaRect != null) {
+      canvas.drawRRect(
+          RRect.fromRectAndRadius(
+              areaRect.inflate(5), const Radius.circular(5)),
+          Paint()
+            ..style = PaintingStyle.stroke
+            ..color = Colors.blue
+            ..strokeWidth = 1 * transform.size
+            ..blendMode = BlendMode.srcOver);
+      canvas.clipRect(areaRect.inflate(5));
+    }
+    if (renderBackground) {
+      cameraViewport.background.build(canvas, size, transform);
+    }
+    if (cameraViewport.bakedElements.isNotEmpty) {
+      var image = cameraViewport.image;
+      var bakedSizeDiff =
+          (transform.size - cameraViewport.scale) / cameraViewport.scale;
+      var pos = transform.globalToLocal(-cameraViewport.toOffset());
+
+      // Draw our baked image, scaling it down with drawImageRect.
+      if (image != null) {
+        canvas.drawImageRect(
+          image,
+          Offset.zero & Size(image.width.toDouble(), image.height.toDouble()),
+          pos &
+              Size(image.width * (1 + bakedSizeDiff),
+                  image.height * (1 + bakedSizeDiff)),
+          Paint(),
+        );
       }
     }
     canvas.saveLayer(Rect.fromLTWH(0, 0, size.width, size.height), Paint());
     canvas.scale(transform.size, transform.size);
     canvas.translate(transform.position.dx, transform.position.dy);
-    document.content
-        .where((element) => !invisibleLayers.contains(element.layer))
-        .toList()
-        .asMap()
-        .forEach((index, element) => paintElement(canvas, element, images));
+    for (var renderer in cameraViewport.unbakedElements) {
+      renderer.build(canvas, size, transform, false);
+    }
     canvas.restore();
   }
 
@@ -163,6 +141,5 @@ class ViewPainter extends CustomPainter {
       document != oldDelegate.document ||
       renderBackground != oldDelegate.renderBackground ||
       transform != oldDelegate.transform ||
-      images != oldDelegate.images ||
-      invisibleLayers != oldDelegate.invisibleLayers;
+      cameraViewport != oldDelegate.cameraViewport;
 }
