@@ -4,7 +4,7 @@ import 'package:phosphor_flutter/phosphor_flutter.dart';
 class OptionButton extends StatelessWidget {
   final Icon icon, selectedIcon;
   final VoidCallback onPressed, onLongPressed;
-  final bool isSelected;
+  final bool selected;
   final String tooltip;
 
   const OptionButton(
@@ -14,7 +14,7 @@ class OptionButton extends StatelessWidget {
       required this.selectedIcon,
       required this.onPressed,
       required this.onLongPressed,
-      this.isSelected = false})
+      this.selected = false})
       : super(key: key);
   @override
   Widget build(BuildContext context) {
@@ -25,24 +25,27 @@ class OptionButton extends StatelessWidget {
         borderRadius: BorderRadius.circular(12),
         onTap: onPressed,
         onLongPress: onLongPressed,
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 8.0, vertical: 3),
-          child: IconTheme(
-              data: Theme.of(context).iconTheme.copyWith(
-                  size: 32,
-                  color: isSelected
-                      ? Theme.of(context).colorScheme.primary
-                      : null),
-              child: Column(
-                children: [
-                  isSelected ? selectedIcon : icon,
-                  if (isSelected)
-                    const Icon(
-                      PhosphorIcons.caretDown,
-                      size: 12,
-                    )
-                ],
-              )),
+        child: AspectRatio(
+          aspectRatio: 1,
+          child: Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 8.0, vertical: 3),
+            child: IconTheme(
+                data: Theme.of(context).iconTheme.copyWith(
+                    size: 28,
+                    color: selected
+                        ? Theme.of(context).colorScheme.primary
+                        : null),
+                child: Column(
+                  children: [
+                    selected ? selectedIcon : icon,
+                    if (selected)
+                      const Icon(
+                        PhosphorIcons.caretDown,
+                        size: 12,
+                      )
+                  ],
+                )),
+          ),
         ),
       ),
     );
