@@ -12,48 +12,10 @@
 module.exports = {
     // By default, Docusaurus generates a sidebar from the docs folder structure
     sidebar: [
-      'intro',
-      'color_picker',
-      'migrating',
-      {
-        type: 'category',
-        label: 'Backgrounds',
-        link: {
-          type: 'doc',
-          id: 'background/intro'
+        {
+          type: 'autogenerated',
+          dirName: '.', // '.' means the current docs folder
         },
-        items: [
-          'background/quad',
-          'background/ruled',
-        ]
-      },
-      {
-        type: 'category',
-        label: 'Painters',
-        link: {
-          type: 'doc',
-          id: 'painters/intro'
-        },
-        items: [
-          'painters/pen',
-          'painters/eraser',
-          'painters/path_eraser',
-          'painters/label',
-          'painters/image',
-          'painters/layer'
-        ]
-      },
-      {
-        type: 'category',
-        label: 'Tools',
-        link: {
-          type: 'doc',
-          id: 'tools/intro',
-        },
-        items: [
-          'tools/hand',
-        ]
-      }
     ],
 
     // But you can create a sidebar manually
